ch3/dependency-parser-nivre/requirements.txt
absl-py==0.1.10